refactor: apply clippy suggestions
Use compound assignment operators, div_ceil, assert!/assert!(!..), unwrap_or_else, 26_usize, size_of_val, and drop an unnecessary borrow. Rename the inner sse2 module to simd to avoid module inception.

+++ src/score.rs
@@ -137,13 +137,13 @@ mod blosum {
         ];
 
         let i = match a.to_ascii_uppercase() {
-            b'*' => 26 as usize,
+            b'*' => 26_usize,
             r if (65..=90).contains(&r) => (r - 65) as usize,
             _ => return None,
         };
 
         let j = match b.to_ascii_uppercase() {
-            b'*' => 26 as usize,
+            b'*' => 26_usize,
             r if (65..=90).contains(&r) => (r - 65) as usize,
             _ => return None,
         };
